+++ upkie/cpp/actuation/tests/BulletInterfaceTest.cpp
@@ -290,24 +290,19 @@ TEST_F(BulletInterfaceTest, ObserveImuOrientation) {
 
 TEST_F(BulletInterfaceTest, MonitorContacts) {
   Dictionary config;
-  config("bullet")("monitor")("contacts")("left_wheel_tire") = true;
-  config("bullet")("monitor")("contacts")("right_wheel_tire") = true;
+  config("bullet")("monitor")("contacts")("wheels")("include")(
+      "left_wheel_tire") = true;
+  config("bullet")("monitor")("contacts")("wheels")("include")(
+      "right_wheel_tire") = true;
   interface_->reset(config);
 
   Dictionary observation;
   interface_->cycle([](const moteus::Output& output) {});
   interface_->observe(observation);
-
   ASSERT_TRUE(observation.has("sim"));
   ASSERT_TRUE(observation("sim").has("contact"));
-  ASSERT_TRUE(observation("sim")("contact").has("left_wheel_tire"));
-  ASSERT_TRUE(observation("sim")("contact").has("right_wheel_tire"));
-  ASSERT_EQ(observation("sim")("contact")("left_wheel_tire")
-                .get<int>("num_contact_points"),
-            0);
-  ASSERT_EQ(observation("sim")("contact")("right_wheel_tire")
-                .get<int>("num_contact_points"),
-            0);
+  ASSERT_TRUE(observation("sim")("contact").has("wheels"));
+  ASSERT_EQ(observation("sim")("contact").get<int>("wheels"), 0);
 }
 
 TEST_F(BulletInterfaceTest, MonitorIMU) {